Created a new enums directory and moved a few files in. Added more documentation to the enums.
