+++ TestRail/Enums/CommandType.cs
@@ -1,0 +1,28 @@
+﻿using EnumStringValues;
+
+namespace TestRail.Enums
+{
+    /// <summary>Command types available</summary>
+    public enum CommandType
+    {
+        /// <summary>For get commands.</summary>
+        [StringValue("get")]
+        Get,
+
+        /// <summary>For add commands.</summary>
+        [StringValue("add")]
+        Add,
+
+        /// <summary>For update commands.</summary>
+        [StringValue("update")]
+        Update,
+
+        /// <summary>For delete commands.</summary>
+        [StringValue("Delete")]
+        Delete,
+
+        /// <summary>For close commands.</summary>
+        [StringValue("close")]
+        Close
+    }
+}
