--- TestRail/Utils/CommandAction.cs
@@ -1,148 +1,0 @@
-﻿using EnumStringValues;
-
-namespace TestRail.Utils
-{
-    /// <summary>
-    /// TODO - Add summary
-    /// </summary>
-    public enum CommandAction
-    {
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("case")]
-        Case,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("Cases")]
-        Cases,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("case_types")]
-        CaseTypes,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("case_fields")]
-        CaseFields,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("milestone")]
-        Milestone,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("milestones")]
-        Milestones,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("plan")]
-        Plan,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("plans")]
-        Plans,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("plan_entry")]
-        PlanEntry,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("project")]
-        Project,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("results")]
-        Results,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("results_for_case")]
-        ResultsForCase,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("results_for_run")]
-        ResultsForRun,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("run")]
-        Run,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("runs")]
-        Runs,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("section")]
-        Section,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("sections")]
-        Sections,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("suite")]
-        Suite,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("suites")]
-        Suites,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("test")]
-        Test,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("tests")]
-        Tests,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("user")]
-        User,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        [StringValue("users")]
-        Users
-    }
-}
