--- TestRail/Utils/CommandType.cs
@@ -1,31 +1,0 @@
-﻿namespace TestRail.Utils
-{
-    /// <summary>Command type's available</summary>
-    public enum CommandType
-    {
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        Get,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        Add,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        Update,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        Delete,
-
-        /// <summary>
-        /// TODO - Add summary
-        /// </summary>
-        Close
-    }
-}
